+++ pyxel/pipelines/processor.py
@@ -14,7 +14,8 @@ from numbers import Number
 
 import numpy as np
 
-from pyxel.detectors import CCD, CMOS
+if t.TYPE_CHECKING:
+    from pyxel.detectors import CCD, CMOS
 from pyxel.evaluator import eval_entry
 from pyxel.pipelines import DetectionPipeline, ModelGroup
 from pyxel.state import get_obj_att
@@ -25,7 +26,7 @@ from pyxel.util.memory import get_size
 class Processor:
     """TBW."""
 
-    def __init__(self, detector: t.Union[CCD, CMOS], pipeline: DetectionPipeline):
+    def __init__(self, detector: t.Union["CCD", "CMOS"], pipeline: DetectionPipeline):
         self._log = logging.getLogger(__name__)
 
         self.detector = detector